requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cymem==2.0.8
docx2txt
idna==2.8
jsonschema==3.0.1
murmurhash==1.0.10
nltk==3.4.5
numpy==1.25.1
pandas==2.1.4
pdfminer.six==20191110
plac==0.9.6
preshed==3.0.9
pycryptodome==3.8.2
pyrsistent==0.15.2
python-dateutil==2.8.2
pytz==2020.1
requests==2.22.0
six==1.12.0
sortedcontainers==2.1.0
spacy==2.3.9
textract==1.6.5
thinc
tqdm==4.38.0
urllib3==1.25.3
wasabi==0.4.0
pdfminer
django
djangorestframework
django-cors-headers
pdfkit
dateparser
gunicorn==20.1